Testing and fixing

+++ tests/test_build_version.py
@@ -136,7 +136,8 @@ class TestVersion(unittest.TestCase):
         """
         Clean all test assets.
         """
-        shutil.rmtree(path=self.TEST_PATH)
+        if self.TEST_PATH.exists():
+            shutil.rmtree(path=self.TEST_PATH)
 
 
 # -->> Execute <<----------------------
